Only delete the build dir in tester recipes if it exists.
The 'rmdir build' step fails if the build dir does not exist, e.g. after a former extract build failure that already nukes the parent directory. Once this state is reached, the bot can't recover anymore.

Review URL: https://codereview.chromium.org/216633004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@261152 0039d316-1c4b-4281-b951-d872f2087c98

+++ recipe_modules/path/api.py
@@ -200,8 +200,9 @@ class PathApi(recipe_api.RecipeApi):
     return self.m.python.inline(
       'rmtree ' + name,
       """
-      import shutil, sys
-      shutil.rmtree(sys.argv[1])
+      import os, shutil, sys
+      if os.path.exists(sys.argv[1]):
+        shutil.rmtree(sys.argv[1])
       """,
       args=[path],
     )
